+++ src/modules/admin/views/hotspot-asset/update.php
@@ -8,7 +8,7 @@
 
 use davidhirtz\yii2\cms\hotspot\modules\admin\controllers\HotspotAssetController;
 use davidhirtz\yii2\cms\hotspot\modules\admin\widgets\forms\HotspotAssetActiveForm;
-use davidhirtz\yii2\cms\modules\admin\widgets\navs\Submenu;
+use davidhirtz\yii2\cms\hotspot\modules\admin\widgets\navs\HotspotSubmenu;
 use davidhirtz\yii2\cms\hotspot\models\HotspotAsset;
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\web\View;
@@ -18,17 +18,11 @@ use davidhirtz\yii2\skeleton\widgets\forms\DeleteActiveForm;
 $this->setTitle(Yii::t('hotspot', 'Edit Hotspot Asset'));
 ?>
 
-<?= Submenu::widget([
-    'model' => $asset->hotspot->asset,
+<?= HotspotSubmenu::widget([
+    'hotspot' => $asset->hotspot,
 ]); ?>
 
-<?php
-$this->setBreadcrumbs([
-    Yii::t('cms', 'Asset') => ['/admin/asset/update', 'id' => $asset->hotspot->asset_id],
-    Yii::t('hotspot', 'Hotspot') => ['/admin/hotspot/update', 'id' => $asset->hotspot->id],
-    Yii::t('media', 'Assets') =>  ['/admin/hotspot/update', 'id' => $asset->hotspot->id, '#' => 'assets'],
-]);
-?>
+<?php $this->setBreadcrumb(Yii::t('cms', 'Asset')); ?>
 
 <?= Html::errorSummary($asset); ?>