+++ src/modules/admin/widgets/panels/HotspotHelpPanel.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace davidhirtz\yii2\cms\hotspot\modules\admin\widgets\panels;
+
+use davidhirtz\yii2\cms\hotspot\models\Hotspot;
+use davidhirtz\yii2\cms\hotspot\modules\admin\controllers\HotspotController;
+use davidhirtz\yii2\cms\modules\admin\widgets\panels\traits\DuplicateButtonTrait;
+use davidhirtz\yii2\skeleton\modules\admin\widgets\panels\HelpPanel;
+
+class HotspotHelpPanel extends HelpPanel
+{
+    use DuplicateButtonTrait;
+
+    public ?Hotspot $model = null;
+
+    public function init(): void
+    {
+        $this->content ??= $this->renderButtonToolbar($this->getButtons());
+        parent::init();
+    }
+
+    /**
+     * @see HotspotController::actionDuplicate()
+     */
+    protected function getButtons(): array
+    {
+        return array_filter([
+            $this->getDuplicateButton(),
+        ]);
+    }
+}